Rimosso orario da Articolo.java + rimossi i rispettivi test

+++ e-shop-manager/src/main/java/it/unipd/mtss/EItem.java
@@ -5,6 +5,6 @@
 package it.unipd.mtss;
 
 public interface EItem {
-	int getPrezzo();
+	Double getPrezzo();
 	String getNome();	
 }
